+++ woocommerce-gateway-ebanx/includes/class-wc-ebanx-account-gateway.php
@@ -31,7 +31,7 @@ class WC_EBANX_Account_Gateway extends WC_EBANX_Redirect_Gateway
      */
     public function is_available()
     {
-        return parent::is_available() && ($this->getTransactionAddress('country') == WC_EBANX_Gateway_Utils::COUNTRY_BRAZIL);
+        return parent::is_available() && $this->getTransactionAddress('country') == WC_EBANX_Gateway_Utils::COUNTRY_BRAZIL;
     }
 
     /**
@@ -66,5 +66,21 @@ class WC_EBANX_Account_Gateway extends WC_EBANX_Redirect_Gateway
             'woocommerce/ebanx/',
             WC_EBANX::get_templates_path()
         );
+    }
+
+    /**
+     * The page of order received, we call them as "Thank you pages"
+     *
+     * @param  WC_Order $order The order created
+     * @return void
+     */
+    public static function thankyou_page($order)
+    {
+        $data = array(
+            'data' => array(),
+            'order_status' => $order->get_status(),
+            'method' => 'account'
+        );
+        parent::thankyou_page($data);
     }
 }